requirements.txt
mcp[cli]==1.10.1
python-dotenv
ipykernel
httpx